+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # 0.7.0 (in progress)
 
+Updated Johann dependency to version 1.0.1.
+
 Added "skip" property to all goals. Plugin executions can now be skipped based on conditional criteria provided by the end user.
+
+Improved error message when a timeout occurs. Previous message printed entire stacktrace of an exception which confused users into thinking it may be a bug in
+the plugin which was usually not the case.
 
 # 0.6.0