+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Updated Johann dependency to version 1.0.1.
 
 Added "skip" property to all goals. Plugin executions can now be skipped based on conditional criteria provided by the end user.
+Users can also provide `dockerCompose.skip` Maven property to skip all plugin executions at once.
 
 Improved error message when a timeout occurs. Previous message printed entire stacktrace of an exception which confused users into thinking it may be a bug in
 the plugin which was usually not the case.